pyproject.toml
[tool.poetry]
name = "ua"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
allure-pytest = "^2.13.5"
requests = "^2.32.3"
pytest-xdist = "^3.6.1"
pytest-rerunfailures = "^15.0"
pydantic = "^2.10.6"
ruff = "^0.9.3"
pre-commit = "^4.1.0"
selenium = "^4.29.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
selenium = "^4.26.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
